requirements.txt
Django==2.1.7
PyJWT==1.7.1
pytz==2018.9
